+++ packages/camera/camera_android/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixes `pausePreview` null pointer error. `pausePreview` should not be called 
   when camera is closed or not configured.
+* Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 0.10.8+13